+++ src/convolutions/sparse_convolution.hpp
@@ -10,39 +10,39 @@ void handle_row_1d_result_colwise_kernel(
     const int &stride_row, const int &stride_col, const int &dilation_row,
     const int &dilation_col, const int &num_snps, const int &num_individuals,
     const int &result_row_size, const int &result_col_size,
-    const vector<vector<int>> &weight, vector<int> &result);
+    const vector<vector<double>> &weight, vector<int> &result);
 
 torch::Tensor sparse_convolution_input_based_optimized(
     const vector<vector<int>> &homo_snps,
     const vector<vector<int>> &hetero_snps, const int &num_snps,
-    const int &num_individuals, const vector<vector<int>> &weight,
+    const int &num_individuals, const vector<vector<double>> &weight,
     const std::tuple<int, int> &stride = std::make_tuple(1, 1),
     const std::tuple<int, int> &dilation = std::make_tuple(1, 1),
-    const int &bias = 0,
+    const double &bias = 0,
     const std::tuple<int, int> &output_size = std::make_tuple(0, 0));
 
 torch::Tensor sparse_convolution_input_based_optimized(
     const vector<vector<int>> &homo_snps,
     const vector<vector<int>> &hetero_snps, const int &num_snps,
-    const int &num_individuals, const vector<vector<int>> &weight,
+    const int &num_individuals, const vector<vector<double>> &weight,
     const int &stride = 1,
     const std::tuple<int, int> &dilation = std::make_tuple(1, 1),
-    const int &bias = 0,
+    const double &bias = 0,
     const std::tuple<int, int> &output_size = std::make_tuple(0, 0));
 
 torch::Tensor sparse_convolution_input_based_optimized(
     const vector<vector<int>> &homo_snps,
     const vector<vector<int>> &hetero_snps, const int &num_snps,
-    const int &num_individuals, const vector<vector<int>> &weight,
+    const int &num_individuals, const vector<vector<double>> &weight,
     const std::tuple<int, int> &stride = std::make_tuple(1, 1),
-    const int &dilation = 1, const int &bias = 0,
+    const int &dilation = 1, const double &bias = 0,
     const std::tuple<int, int> &output_size = std::make_tuple(0, 0));
 
 torch::Tensor sparse_convolution_input_based_optimized(
     const vector<vector<int>> &homo_snps,
     const vector<vector<int>> &hetero_snps, const int &num_snps,
-    const int &num_individuals, const vector<vector<int>> &weight,
-    const int &stride = 1, const int &dilation = 1, const int &bias = 0,
+    const int &num_individuals, const vector<vector<double>> &weight,
+    const int &stride = 1, const int &dilation = 1, const double &bias = 0,
     const std::tuple<int, int> &output_size = std::make_tuple(0, 0));
 
 void handle_row_1d_result_colwise(
